requirements.txt
# WatchTower SIEM - Python Dependencies
# Pin minor versions for reproducible builds

# ── Web framework ────────────────────────────────────────────────────────────
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==22.0.0
gevent==24.11.1

# ── Flask extensions ─────────────────────────────────────────────────────────
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Mail==0.10.0
Flask-WTF==1.2.1
Flask-Compress==1.15
Flask-Cors==4.0.1

# ── Database ─────────────────────────────────────────────────────────────────
pymongo[srv]==4.7.3
dnspython==2.6.1

# ── Task queue ───────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.0.7
kombu==5.3.7
flower==2.0.1

# ── Auth / Security ──────────────────────────────────────────────────────────
bcrypt==4.1.3
PyJWT==2.8.0
pyotp==2.9.0
qrcode[pil]==7.4.2
bleach==6.1.0
cryptography==42.0.8

# ── AI providers ─────────────────────────────────────────────────────────────
anthropic==0.31.1
openai==1.40.0

# ── Validation / Serialization ───────────────────────────────────────────────
marshmallow==3.21.3
python-dateutil==2.9.0
jsonschema==4.23.0

# ── Monitoring / Observability ───────────────────────────────────────────────
prometheus-flask-exporter==0.23.1
structlog==24.4.0
sentry-sdk[flask]==2.13.0

# ── HTTP client ──────────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.27.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
Pillow==10.4.0

# ── Optional: GeoIP ──────────────────────────────────────────────────────────
# maxminddb==2.6.2
# geoip2==4.8.0

# ── PDF export ──────────────────────────────────────────────────────────────
reportlab==4.4.1

# ── Sigma rule import ────────────────────────────────────────────────────────
pyyaml==6.0.2
